fix: modal size on desktop
modals are surprisingly more pleasant to work with now in chakra v1.
the size "5xl" adjusts accordingly to the desktop width so we dont need
to supply it with a responsive array anymore.

notable annoyance: when clicking the footer modals repetitively, the
hackacomm toast keeps popping up on each route change with the hash #