fix: slight anti-pattern to override a chakra default for toasts
chakra's toast has a parent component called alert which applies 0.5rem
margin to it. i decided to override the bottom margin with 0 and using
the css important flag. i believe the only other place we are using a
toast is for the email ux. which comes from the top so i felt that it
was safe to remove the bottom margin.