fix: avoid toast constantly popping up annoyance
previously, the toast would pop up when a modal is opened and again when
the modal is closed. now it will not pop up if the user has previously
dismissed it manually already and checks the footer modals.
the toast will come back tho if the user makes a full page change.